fix: 이미지 반영 안되는 문제
